Works on readthedocs!
I understand why it was failing locally (see previous git commit
message). I was using `python setup.py install` on a virtual
environment. In this environment, I had setup the imports in the way I
wanted, but the virtual environment did not have sphinx installed. So it
was calling sphinx from my actual python distribution which was not
looking into the virtual environment.

Hence, things work correctly on readthedocs.